Minor refactors and UI updates
Changed all indices to long
Added menu to LogFragment
Added "Clear old logs"
Displayed duration nicely in Logs and Notification
Only showing logs >30 second duration

+++ app/src/main/java/io/github/mikecornflake/apptimelimiter/database/dao/PackageDao.kt
@@ -25,7 +25,7 @@ interface PackageDao {
     fun getAllPackages(): Flow<List<Package>>
 
     @Query("SELECT * FROM package WHERE packageId = :packageId")
-    fun getPackage(packageId: Int): Flow<Package>
+    fun getPackage(packageId: Long): Flow<Package>
 
     @Query("SELECT * FROM package WHERE packageName = :packageName")
     fun getPackage(packageName: String): Flow<Package>
